pyproject.toml
[project]
name = "humanand"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyrefly>=0.53.0",
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "anthropic>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
